added Header component to Traffic Page
