+++ App/app/traffic/page.styles.ts
@@ -3,8 +3,8 @@ import styled from "styled-components";
 const TrafficPage = styled.div`
   background: #16151c;
   color: #fff;
-  height: 100%;
-  padding: 46px 20px 24px 45px;
+  width: 100%;
+  padding: 0 10px 24px 21px;
 `;
 
 export { TrafficPage };
