+++ App/components/dashboard/bottomSection/BottomSection.styles.ts
