+++ App/components/dashboard/midSection/MidSection.styles.ts
